+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `biurad/cache-manager` will be documented in this file.
 
+## 0.1.7 - 2019-12-17
+
+- Removed `CacheResolver` class
+- Removed *"mongodb"* cache driver, since it wouldn't be used
+- Added `CacheBridge` class for better performance
+- Added `Connection` class for easy connection to servers
+- Fixed Security issues regarding caching.
+- Updated ReadMe.md file
+
 ## 0.1.5 - 2019-11-21
 
 - Fixed issues with phpunit test